+++ src/reanimated2/animation/decay/utils.ts
@@ -5,15 +5,13 @@ import type {
   Animation,
   ReduceMotion,
   Timestamp,
+  RequiredKeys,
 } from '../../../reanimated2/commonTypes';
 import { isWeb } from '../../PlatformChecker';
 
 const IS_WEB = isWeb();
 export const VELOCITY_EPS = IS_WEB ? 1 / 20 : 1;
 export const SLOPE_FACTOR = 0.1;
-
-export type DecayConfig = Partial<Omit<DefaultDecayConfig, 'rubberBandFactor'>>;
-export type WithDecayConfig = DecayConfig;
 
 export interface DecayAnimation extends Animation<DecayAnimation> {
   lastTimestamp: Timestamp;
@@ -32,35 +30,26 @@ export interface InnerDecayAnimation
 export interface DefaultDecayConfig {
   deceleration: number;
   velocityFactor: number;
-  clamp?: number[];
+  clamp?: [min: number, max: number];
   velocity: number;
   reduceMotion?: ReduceMotion;
   rubberBandEffect?: boolean;
   rubberBandFactor: number;
 }
 
-export function validateConfig(config: DefaultDecayConfig): void {
+// If user wants to use rubber band decay animation we have to make sure he has provided clamp
+export type RubberBandDecayConfig = RequiredKeys<
+  DefaultDecayConfig,
+  'clamp'
+> & { rubberBandEffect: true };
+
+export function isValidRubberBandConfig(
+  config: DefaultDecayConfig
+): config is RubberBandDecayConfig {
   'worklet';
-  if (config.clamp) {
-    if (!Array.isArray(config.clamp)) {
-      throw new Error(
-        `[Reanimated] \`config.clamp\` must be an array but is ${typeof config.clamp}.`
-      );
-    }
-    if (config.clamp.length !== 2) {
-      throw new Error(
-        `[Reanimated] \`clamp array\` must contain 2 items but is given ${config.clamp.length}.`
-      );
-    }
-  }
-  if (config.velocityFactor <= 0) {
-    throw new Error(
-      `[Reanimated] \`config.velocityFactor\` must be greather then 0 but is ${config.velocityFactor}.`
-    );
-  }
-  if (config.rubberBandEffect && !config.clamp) {
-    throw new Error(
-      '[Reanimated] You need to set `clamp` property when using `rubberBandEffect`.'
-    );
-  }
+  return (
+    Array.isArray(config.clamp) &&
+    config.clamp.length === 2 &&
+    !!config.rubberBandEffect
+  );
 }
